# agent/api/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing btrfs volumes, snapshots, clones, NFS exports, and background tasks.",
        "title": "btrfs-nfs-csi Agent API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/clones": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new volume from a snapshot. Returns existing on conflict.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clones"
                ],
                "summary": "Create a clone from snapshot",
                "parameters": [
                    {
                        "description": "Clone configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CloneCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Volume already exists",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/volumes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all volumes. Use detail=true for full metadata. Supports pagination and label filtering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "List volumes",
                "parameters": [
                    {
                        "enum": [
                            "true"
                        ],
                        "type": "string",
                        "description": "Return full detail",
                        "name": "detail",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor from previous response",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Label filter (key=value), repeatable",
                        "name": "label",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new btrfs subvolume. Returns the existing volume on conflict.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Create a volume",
                "parameters": [
                    {
                        "description": "Volume configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VolumeCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Volume already exists",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    }
                }
            }
        },
        "/v1/volumes/clone": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new volume from another volume (internal snapshot + clone). Returns existing on conflict.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clones"
                ],
                "summary": "Clone a volume",
                "parameters": [
                    {
                        "description": "Clone configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VolumeCloneRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Volume already exists",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    }
                }
            }
        },
        "/v1/volumes/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns detailed metadata for a single volume.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Get a volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a volume and its btrfs subvolume. Fails if the volume has active exports.",
                "tags": [
                    "volumes"
                ],
                "summary": "Delete a volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "423": {
                        "description": "Volume has active exports",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Patches volume properties. Nil/omitted fields are left unchanged.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Update a volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VolumeUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VolumeDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CloneCreateRequest": {
            "type": "object",
            "properties": {
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "^[a-zA-Z0-9_-]{1,128}$",
                    "type": "string"
                },
                "snapshot": {
                    "description": "source snapshot name",
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "models.ExportDetailResponse": {
            "type": "object",
            "properties": {
                "client": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.VolumeCloneRequest": {
            "type": "object",
            "properties": {
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "^[a-zA-Z0-9_-]{1,128}$",
                    "type": "string"
                },
                "source": {
                    "description": "source volume name",
                    "type": "string"
                }
            }
        },
        "models.VolumeCreateRequest": {
            "type": "object",
            "properties": {
                "compression": {
                    "description": "\"zstd\", \"zstd:3\", \"zlib\", \"zlib:5\", \"lzo\", or \"\"",
                    "type": "string"
                },
                "gid": {
                    "description": "owner GID (0-65534)",
                    "type": "integer"
                },
                "labels": {
                    "description": "key: ^[a-z0-9][a-z0-9._-]{0,62}$, value: ^[a-zA-Z0-9._-]{0,128}$",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mode": {
                    "description": "octal permission string, e.g. \"0755\" (max 7777)",
                    "type": "string"
                },
                "name": {
                    "description": "^[a-zA-Z0-9_-]{1,128}$",
                    "type": "string"
                },
                "nocow": {
                    "description": "disable copy-on-write (chattr +C)",
                    "type": "boolean"
                },
                "quota_bytes": {
                    "description": "btrfs qgroup limit (0 = no quota)",
                    "type": "integer"
                },
                "size_bytes": {
                    "description": "subvolume size in bytes",
                    "type": "integer"
                },
                "uid": {
                    "description": "owner UID (0-65534)",
                    "type": "integer"
                }
            }
        },
        "models.VolumeDetailResponse": {
            "type": "object",
            "properties": {
                "clients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExportDetailResponse"
                    }
                },
                "compression": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "gid": {
                    "type": "integer"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "last_attach_at": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nocow": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "quota_bytes": {
                    "type": "integer"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "uid": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "used_bytes": {
                    "type": "integer"
                }
            }
        },
...
        "models.VolumeListResponse": {
            "type": "object",
            "properties": {
                "next": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.VolumeResponse"
                    }
                }
            }
        },
        "models.VolumeResponse": {
            "type": "object",
            "properties": {
                "clients": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "used_bytes": {
                    "type": "integer"
                }
            }
        },
        "models.VolumeUpdateRequest": {
            "type": "object",
            "properties": {
                "compression": {
                    "description": "see VolumeCreateRequest.Compression",
                    "type": "string"
                },
                "gid": {
                    "description": "new owner GID (0-65534)",
                    "type": "integer"
                },
                "labels": {
                    "description": "replaces all labels",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mode": {
                    "description": "octal permission string (max 7777)",
                    "type": "string"
                },
                "nocow": {
                    "description": "toggle copy-on-write",
                    "type": "boolean"
                },
                "size_bytes": {
                    "description": "new size (must be \u003e= current)",
                    "type": "integer"
                },
                "uid": {
                    "description": "new owner UID (0-65534)",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Tenant token as \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}